Working dir for a receipt
In some cases, it's desirable for a container to use working dir defined in Docker image metadata.

However, the current code always sets working dir to the current directory, thus overriding the value of working dir.

To address this situation and to maintain backward compatibility with existing clients, it seems reasonable to introduce additional fields to the structures Receipt and Config correspondingly.

Proposed changes would allow a client to explicitly decide whether to use working dir defined in image metadata or the current directory.

Although this solution might not be ideal in terms of client expectations, it could be considered as an intermediate one without breaking API changes.